+++ R/predictions.R
@@ -29,9 +29,12 @@ select_prediction_method <- function(fun,
   } else if (fun == "brmsfit") {
     # brms-objects -----
     fitfram <- get_predictions_stan(model, expanded_frame, ci.lvl, type, faminfo, ppd, ...)
-  } else if (fun == "coxph") {
+  } else if (fun == "coxph" && type != "surv" && type != "cumhaz") {
     # coxph-objects -----
     fitfram <- get_predictions_coxph(model, expanded_frame, ci.lvl, ...)
+  } else if (fun == "coxph" && type %in% c("surv", "cumhaz")) {
+    # coxph-objects -----
+    fitfram <- get_predictions_survival(model, expanded_frame, ci.lvl, type, terms, ...)
   } else if (fun == "lrm") {
     # lrm-objects -----
     fitfram <- get_predictions_lrm(model, expanded_frame, ci.lvl, linv, ...)
@@ -1019,6 +1022,70 @@ get_predictions_coxph <- function(model, fitfram, ci.lvl, ...) {
   fitfram
 }
 
+
+# predictions for coxph, survival and cumulative hazard ----
+
+get_predictions_survival <- function(model, fitfram, ci.lvl, type, terms, ...) {
+  # does user want standard errors?
+  se <- !is.null(ci.lvl) && !is.na(ci.lvl)
+
+  # compute ci, two-ways
+  if (!is.null(ci.lvl) && !is.na(ci.lvl))
+    ci <- 1 - ((1 - ci.lvl) / 2)
+  else
+    ci <- .975
+
+  if (!requireNamespace("survival", quietly = TRUE)) {
+    stop("Package `survival` required. Please install it.", call. = FALSE)
+  }
+
+  # get survial probabilities and cumulative hazards
+
+  prdat <- survival::survfit(
+    model,
+    newdata = fitfram,
+    se.fit = TRUE,
+    conf.int = ci,
+    ...
+  )
+
+  # check what user requested and either returnd surv probs
+  # or cumulative hazards, including CI
+
+  if (type == "surv") {
+    pr <- prdat$surv
+    lower <- prdat$lower
+    upper = prdat$upper
+  } else {
+    pr <- prdat$cumhaz
+    lower <- pr - stats::qnorm(ci) * prdat$std.err
+    upper <- pr + stats::qnorm(ci) * prdat$std.err
+    # ugly fix...
+    pr[which(pr < 0)] <- 0
+    lower[which(lower < 0)] <- 0
+    upper[which(upper < 0)] <- 0
+  }
+
+  # Now we need the groups, as survfit() only returns numeric indices
+
+  clean_terms <- get_clear_vars(terms)
+  ff <- fitfram[clean_terms]
+
+  purrr::map_df(sjmisc::seq_row(ff), function(i) {
+
+    dat <- data.frame(
+      time = prdat$time,
+      predicted = pr[, i],
+      conf.low = lower[, i],
+      conf.high = upper[, i]
+    )
+
+    dat2 <- dplyr::bind_cols(purrr::map(sjmisc::seq_col(ff), ~ ff[i, .x]))
+    colnames(dat2) <- clean_terms
+
+    cbind(dat[, 1, drop = FALSE], dat2, dat[, 2:4])
+  })
+}
 
 
 # predictions for gam ----