+++ R/utils_expand_data_to_grid.R
@@ -140,7 +140,7 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, pre
   # level for factors and most common element for character vectors
 
   if (fac.typical) {
-    const.values <- lapply(mf[, alle, drop = FALSE], function(x) sjstats::typical_value(x, typ.fun, weight.by = w))
+    const.values <- lapply(mf[, alle, drop = FALSE], function(x) sjstats::typical_value(x, fun = typ.fun, weights = w))
   } else {
     # if factors should not be held constant (needed when computing
     # std.error for merMod objects), we need all factor levels,
@@ -150,7 +150,7 @@ get_expanded_data <- function(model, mf, terms, typ.fun, fac.typical = TRUE, pre
         if (is.factor(x))
           levels(x)
         else
-          sjstats::typical_value(x, typ.fun, weight.by = w)
+          sjstats::typical_value(x, fun = typ.fun, weights = w)
       })
   }